Agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
pyodbc==5.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
